requirements.txt
pymongo==4.7.1
google-cloud-storage==2.16.0
IP2Location==8.10.0
requests==2.32.3
beautifulsoup4==4.12.3
tqdm==4.66.4
python-dotenv==1.0.1
